re #32, re #59 refactoring out epics client into separate files.
